+++ src/utilities/SpacingDemo.js
@@ -10,7 +10,7 @@ export const SpacingDemo = () => {
         <div className="p-grid spacing-demo">
             <div className="p-col-12">
                 <div className="card">
-                    <h3>Spacing</h3>
+                    <h4>Spacing</h4>
                     <p><Link to="grid">PrimeFlex</Link> provides various spacing utilities to modify
                     an element's layout.</p>
 
@@ -62,15 +62,15 @@ export const SpacingDemo = () => {
 
             <div className="p-col-12 p-md-6">
                 <div className="card">
-                    <h5>Use Case 1</h5>
+                    <h4>Use Case 1</h4>
                     <p>Spacing is a handy utility when elements wrap when screen size get smaller. In example below, when buttons wrap
                     they are displayed on top of each other whereas with spacing this issue can easily be resolved.</p>
 
-                    <h6>Without Spacing</h6>
+                    <h5>Without Spacing</h5>
                     <Button type="button" label="Button 1" style={{ width: '20rem' }} />
                     <Button type="button" label="Button 2" style={{ width: '20rem' }} />
 
-                    <h6>With Spacing</h6>
+                    <h5>With Spacing</h5>
                     <Button type="button" label="Button 1" style={{ width: '20rem' }} className="p-mr-2 p-mb-2" />
                     <Button type="button" label="Button 2" style={{ width: '20rem' }} className="p-mb-2" />
                 </div>
@@ -78,11 +78,11 @@ export const SpacingDemo = () => {
 
             <div className="p-col-12 p-md-6">
                 <div className="card">
-                    <h5>Use Case 2</h5>
+                    <h4>Use Case 2</h4>
                     <p>Spacing utilities also work seamlessly with responsive PrimeFlex grid utilty. Example below demonstrates a case where on a smaller screen inputs receive a margin to position themselves
                     separately in stacked mode, reduce the dimension of the window to view the difference.</p>
 
-                    <h6>Without Spacing</h6>
+                    <h5>Without Spacing</h5>
                     <div className="p-grid p-formgrid p-fluid">
                         <div className="p-col-12 p-lg-4">
                             <InputText />
@@ -95,7 +95,7 @@ export const SpacingDemo = () => {
                         </div>
                     </div>
 
-                    <h6>With Spacing</h6>
+                    <h5>With Spacing</h5>
                     <div className="p-grid p-formgrid p-fluid">
                         <div className="p-col-12 p-mb-2 p-lg-4 p-mb-lg-0">
                             <InputText />